Move and rename modules to a more self descriptive arrangement
