+++ app/_lib/fitbit/messaging.js
