+++ app/_lib/fitbit/settings.js
