+++ app/ui/index.js
@@ -1,7 +1,7 @@
 import document from "document";
-import * as messaging from "app/fitbit/messaging";
-import * as vibration from "app/ui/vibration";
-import {nextTile} from "app/ui/tiles";
+import * as messaging from "../fitbit/messaging";
+import * as vibration from "./vibration";
+import {nextTile} from "./tiles";
 
 const icons = {
   notConnected: 80,
@@ -16,20 +16,6 @@ const setCompanionIcon = currentState => {
   iconMask.x = icons[currentState.companionConnect];
 };
 
-const setConsoleText = currentState => {
-  const head = document.getElementById("console-head");
-  const body = document.getElementById("console-body");
-
-  // only change if the value is not null
-  // set to "" to clear a field instead of null
-  if (currentState.console.headText !== null) {
-    head.text = currentState.console.headText;
-  }
-  if (currentState.console.bodyText!== null) {
-    body.text = currentState.console.bodyText;
-  }
-};
-
 const setDemoVisible = currentState => {
   const demoIcon = document.getElementById("demo");
   demoIcon.style.display = currentState.isDemo
@@ -39,7 +25,6 @@ const setDemoVisible = currentState => {
 
 export const updateUI = currentState => {
   setCompanionIcon(currentState);
-  setConsoleText(currentState);
   setDemoVisible(currentState);
 };
