+++ companion/_lib/fitbit/messaging.js
