+++ companion/_lib/fitbit/settings.js
