+++ companion/_lib/utils.js
